requirements.txt
streamlit~=1.14.0
Pillow~=9.3.0
seaborn-image~=0.5.0
Pillow~=9.3.0
tqdm~=4.64.1